views: Add backward compatibility for the send-error-report
Newer versions of django handle application/json in a more correct way
by having the payload in the request body rather than a parameter to the
HTTP post request. So that we can continue to parse data correctly from
old send-error-report scripts we need to undo the url encoding of the
data.

see https://docs.djangoproject.com/en/dev/releases/1.5/
section #non-form-data-in-http-requests

Signed-off-by: Michael Wood <michael.g.wood@intel.com>