views/tests: Add a user-agent path to switch payload decode paths
To enable us to drop support for the URL encoded json payload and use
the more conventional non-encoded application/json payload we add a
switch on the client's http user agent, which will allow us to identify which
type of payload is being submitted and pave the way for future changes.

Also add tests to test this new path. Refactor tests so that they delete
the data after running to enable retrieval without having to know the PK

[YOCTO #7245]

Signed-off-by: Michael Wood <michael.g.wood@intel.com>