Add a similar errors matching mechanism
This adds the feature to recognise if an error has been reported before.
It matches it against the distance with an offset of bytes by using the
first error in the database as the prime error. The accuracy is
adjustable by setting SIMILAR_FAILURE_DISTANCE. If using the JSON
response to a submission you will also get the number of the duplicates
and a link to view them.

[YOCTO #7798]

Signed-off-by: Michael Wood <michael.g.wood@intel.com>